+++ skellycam/core/frames/wrangling/frame_listener_process.py
@@ -81,7 +81,7 @@ class FrameListenerProcess:
             logger.info(f"Frame exporter process received KeyboardInterrupt, shutting down gracefully...")
         finally:
             logger.trace(f"Stopped listening for multi-frames")
-            if not dto.ipc_flags.kill_camera_group_flag.value:
+            if not dto.ipc_flags.kill_camera_group_flag.value and not dto.ipc_flags.global_kill_flag.value:
                 raise ValueError("FrameListenerProcess was closed before the camera group kill flag was set.")
 
     def is_alive(self) -> bool:
