+++ skellycam/core/frames/wrangling/frame_router_process.py
@@ -99,7 +99,7 @@ class FrameRouterProcess:
             logger.info(f"Frame exporter process received KeyboardInterrupt, shutting down gracefully...")
         finally:
             logger.trace(f"Stopped listening for multi-frames")
-            if not dto.ipc_flags.kill_camera_group_flag.value:
+            if not dto.ipc_flags.kill_camera_group_flag.value and not dto.ipc_flags.global_kill_flag.value:
                 raise ValueError("FrameRouter should only be closed after global kill flag is set")
 
             if video_recorder_manager:
